Created routes to index, register, login and links on pages (did some styling to make it easier on the eyes for development)
