+++ views/index.ejs
@@ -1,6 +1,10 @@
 <%- include('partials/head.ejs') -%>
+<%- include('partials/navbar.ejs') -%>
 
-<title>test</title>
-<h1>Index page</h1>
+<title>SquareHelp - index.html</title>
+
+<div class="wrapper">
+  <h1>Index page</h1>
+</div>
 
 <%- include('partials/footer.ejs') -%>
