+++ views/login.ejs
@@ -1,10 +1,10 @@
 <%- include('partials/head.ejs') -%>
 <%- include('partials/navbar.ejs') -%>
 
-<title>SquareHelp - error.html</title>
+<title>SquareHelp - login.html</title>
 
 <div class="wrapper">
-  <h1>Error Page</h1>
+  <h1>Login page</h1>
 </div>
 
 <%- include('partials/footer.ejs') -%>
