+++ views/partials/footer.ejs
@@ -1,4 +1,3 @@
 <footer>
 <script src="js/materialize.min.js"></script>
-
 </footer>
