+++ views/partials/head.ejs
@@ -7,4 +7,4 @@
     <link href="css/style.css" rel="stylesheet">
   </head>
   <body>
-  <h2>Head.ejs</h2>
+  <!-- head ejs is loaded -->
