+++ views/partials/navbar.ejs
@@ -1,0 +1,11 @@
+  <div class="navbar-fixed">
+    <nav class="grey darken-4">
+      <div class="nav-wrapper">
+        <a href="/" class="brand-logo">SquareHelp</a>
+        <ul class="right hide-on-med-and-down">
+          <li><a href="login">Login</a></li>
+          <li><a href="register">Register</a></li>
+        </ul>
+      </div>
+    </nav>
+  </div>
